test: split runtime CLI integration scenarios
