--- packages/cli/test/runtime-cli.integration.test.ts
@@ -1,1751 +1,0 @@
-// harness-test-tier: integration
-import assert from "node:assert/strict";
-import { spawn, spawnSync } from "node:child_process";
-import { randomUUID } from "node:crypto";
-import {
-  existsSync,
-  mkdirSync,
-  mkdtempSync,
-  readFileSync,
-  readdirSync,
-  realpathSync,
-  rmSync,
-  writeFileSync,
-} from "node:fs";
-import { tmpdir } from "node:os";
-import path from "node:path";
-import test from "node:test";
-import { makeTaskEventReader } from "../../kernel/src/index.ts";
-import { safePath } from "../../daemon/src/protocol/daemon-protocol.contract.ts";
-import { localUserDaemonEndpoint, runCommandThroughDaemon } from "../src/daemon/client.ts";
-import { writeProviderExecutable } from "../../daemon/test/fixtures/runtime-stub.ts";
-import { realizedTaskPlan as realizedPlan } from "../../../tools/fixtures/task-plan.mjs";
-
-const cli = path.resolve("packages/cli/src/index.ts");
-
-test("real CLI runs, archives task-bound dispatches, resumes, waits through status, streams, and cancels idempotently", async (context) => {
-  const parent = mkdtempSync(path.join(process.platform === "win32" ? tmpdir() : "/tmp", "ha-runtime-cli-")),
-    root = path.join(parent, "repo"),
-    userRoot = path.join(parent, "user"),
-    daemonId = `runtime-cli-test-${randomUUID()}`,
-    binRoot = path.join(parent, "bin"),
-    tracker = path.join(root, ".batch-tracker"),
-    nonzeroTrace = path.join(root, ".notify-nonzero.json"),
-    notificationStarted = path.join(root, ".notify-hold-started"),
-    notificationRelease = path.join(root, ".notify-hold-release"),
-    notificationFinished = path.join(root, ".notify-hold-finished"),
-    notificationOnce = path.join(root, ".notify-once"),
-    version = "0.0.0-runtime-cli-fixture",
-    {
-      HARNESS_DAEMON_ENDPOINT: _endpoint,
-      HARNESS_DAEMON_REPO_ID: _repoId,
-      HARNESS_DAEMON_ID: _daemonId,
-      ...baseEnv
-    } = process.env;
-  mkdirSync(root, { recursive: true });
-  mkdirSync(binRoot, { recursive: true });
-  writeProgressProvider(path.join(binRoot, "codex"), version);
-  writeProviderExecutable(
-    path.join(binRoot, "gh"),
-    'if (process.argv[2] !== "run" || process.argv[3] !== "list") process.exit(1); console.log("[]");\n',
-  );
-  const nonzeroNotifier = writeProviderExecutable(
-      path.join(parent, "notify-nonzero"),
-      `const fs = require("node:fs"), payload = JSON.parse(fs.readFileSync(0, "utf8")); fs.writeFileSync(".notify-nonzero.json", JSON.stringify({ cwd: process.cwd(), environment: process.env, payload })); process.exit(23);\n`,
-    ),
-    holdNotifier = writeProviderExecutable(
-      path.join(parent, "notify-hold"),
-      'const fs = require("node:fs"); fs.readFileSync(0, "utf8"); ' +
-        'const watcher = fs.watch(".", (_, filename) => { if (filename !== ".notify-hold-release") return; ' +
-        'watcher.close(); fs.writeFileSync(".notify-hold-finished", "finished\\n"); }); ' +
-        'fs.writeFileSync(".notify-hold-started", "started\\n");\n',
-    ),
-    onceNotifier = writeProviderExecutable(
-      path.join(parent, "notify-once"),
-      `const fs = require("node:fs"); fs.readFileSync(0, "utf8"); fs.appendFileSync(".notify-once", "once\\n");\n`,
-    );
-  const env = {
-    ...baseEnv,
-    HOME: path.join(parent, "home"),
-    TMPDIR: process.platform === "win32" ? baseEnv.TMPDIR : "/tmp",
-    PATH: [
-      binRoot,
-      ...(process.env.PATH ?? "")
-        .split(path.delimiter)
-        .filter((entry) => ["codex", "codex.cmd", "codex.exe"].every((name) => !existsSync(path.join(entry, name)))),
-    ].join(path.delimiter),
-    OPENAI_API_KEY: "sk-must-not-reach-notifier",
-    HARNESS_NOTIFY_TEST_SECRET: "must-not-reach-notifier",
-    HARNESS_DAEMON_USER_ROOT: userRoot,
-    HARNESS_DAEMON_ID: daemonId,
-    HARNESS_DAEMON_ENDPOINT:
-      process.platform === "win32"
-        ? localUserDaemonEndpoint(userRoot, daemonId)
-        : path.join("/tmp/harness-anything", path.basename(localUserDaemonEndpoint(userRoot, daemonId))),
-    HARNESS_ACTOR: "agent:runtime-cli-test",
-  };
-  try {
-    assert.equal(run(root, env, ["daemon", "start", "--service"]).ok, true);
-    run(root, env, ["init", "--repo-id", "runtime-cli", "--person-id", "owner", "--display-name", "Owner"]);
-    mkdirSync(path.join(root, "harness", "skills", "review"), { recursive: true });
-    writeFileSync(
-      path.join(root, "harness", "skills", "review", "SKILL.md"),
-      "---\nname: review\ndescription: Review\n---\nReview fixture.\n",
-    );
-    const identities = [
-        { id: "fable", name: "Fable", instructions: "Lead precisely.", runtime_type: "codex", role: "commander" },
-        { id: "terra", name: "Terra", instructions: "Review precisely.", runtime_type: "codex", role: "worker" },
-        {
-          id: "outsider",
-          name: "Outsider",
-          instructions: "Work outside the squad.",
-          runtime_type: "codex",
-          role: "worker",
-        },
-        {
-          id: "opencode-worker",
-          name: "OpenCode Worker",
-          instructions: "Use OpenCode.",
-          runtime_type: "opencode",
-          role: "worker",
-        },
-        {
-          id: "any-worker",
-          name: "Any Worker",
-          instructions: "Use any compatible runtime.",
-          runtime_type: "any",
-          role: "worker",
-        },
-      ],
-      squadSource = path.join(parent, "core-squad");
-    for (const identity of identities) {
-      const agentSource = path.join(parent, identity.id);
-      writeIdentity(agentSource, {
-        id: identity.id,
-        title: identity.name,
-        kind: "agent",
-        agent: {
-          ...identity,
-          skills: [{ id: "review", path: "skills/review" }],
-          prompts: ["prompt://review"],
-          preset: "standard-task",
-        },
-      });
-      run(root, env, ["agent", "install", "--source", agentSource]);
-    }
-    writeIdentity(squadSource, {
-      id: "core-squad",
-      title: "Core Squad",
-      kind: "squad",
-      squad: {
-        id: "core-squad",
-        name: "Core Squad",
-        leader: "fable",
-        workers: ["terra", "opencode-worker"],
-        leaderTurnBudget: 8,
-        roster: "# Core Squad\n\nFable delegates to Terra and OpenCode Worker.",
-      },
-    });
-    run(root, env, ["squad", "install", "--source", squadSource]);
-    const agents = JSON.parse(String(run(root, env, ["agent", "list"]).evidence)) as { agents: Array<{ id: string }> },
-      squads = JSON.parse(String(run(root, env, ["squad", "list"]).evidence)) as { squads: Array<{ id: string }> };
-    assert.deepEqual(
-      agents.agents.map(({ id }) => id),
-      ["any-worker", "fable", "opencode-worker", "outsider", "terra"],
-    );
-    assert.deepEqual(
-      squads.squads.map(({ id }) => id),
-      ["core-squad"],
-    );
-    writeIdentity(squadSource, {
-      id: "core-squad",
-      title: "Core Squad",
-      kind: "squad",
-      squad: {
-        id: "core-squad",
-        name: "Core Squad",
-        leader: "fable",
-        workers: ["terra", "opencode-worker"],
-        leaderTurnBudget: 8,
-        roster: "# Core Squad\n\nFable delegates; humans edited this roster.",
-      },
-    });
-    run(root, env, ["squad", "install", "--source", squadSource]);
-    const squad = JSON.parse(String(run(root, env, ["squad", "inspect", "core-squad"]).evidence)) as {
-      squad: { roster: string };
-    };
-    assert.match(squad.squad.roster, /humans edited/u);
-    const taskId = "task-runtime-archive",
-      executionId = "exec-runtime-archive",
-      created = run(root, env, ["task", "create", "--id", taskId, "--admin", "--title", "Runtime archive"]),
-      packagePath = String(created.packagePath),
-      artifactRoot = path.join(root, "harness", packagePath, "artifacts");
-    const planPath = path.join(root, "harness", packagePath, "task_plan.md");
-    published(root, env, created);
-    writeFileSync(planPath, realizedPlan("Runtime archive"));
-    run(root, env, ["doc", "sync", "--submit", "--path", `${packagePath}/task_plan.md`]);
-    const inventory = run(root, env, ["runtime", "instance", "list"]),
-      installation = (inventory.installations as Array<Record<string, unknown>>).find(
-        (row) => row.kindId === "codex" && row.version === `codex ${version}`,
-      );
-    assert.ok(installation, JSON.stringify(inventory));
-    run(root, env, [
-      "runtime",
-      "instance",
-      "create",
-      "--id",
-      "cli-worker",
-      "--name",
-      "CLI Worker",
-      "--kind",
-      "codex",
-      "--provider",
-      "openai",
-      "--model",
-      "runtime-test-model",
-      "--auth",
-      "subscription",
-    ]);
-    const missingInstance = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "missing-instance",
-      "--prompt",
-      "must reject",
-      "--detach",
-    ]);
-    assert.equal(missingInstance.status, 1);
-    assert.equal(missingInstance.receipt.code, "runtime_instance_not_found");
-    assert.equal(missingInstance.receipt.dispatchId, undefined);
-    const directPrompt = run(root, env, ["runtime", "run", "cli-worker", "--prompt", "file prompt", "--no-stream"]);
-    assert.equal((directPrompt.result as Record<string, unknown>).text, "final:file prompt");
-    for (const directory of ["missions", "dispatches", "reports"])
-      assert.equal(existsSync(path.join(artifactRoot, directory)), false, `${directory} must not exist without --task`);
-    const automaticTaskId = `${taskId}-automatic`;
-    const automaticTask = run(root, env, [
-      "task",
-      "create",
-      "--id",
-      automaticTaskId,
-      "--admin",
-      "--title",
-      "Automatic lease",
-    ]);
-    const automaticPackagePath = String(automaticTask.packagePath),
-      automaticPlanPath = `${automaticPackagePath}/task_plan.md`,
-      oneSectionMissing = realizedPlan("Automatic lease").replace(
-        /\n\n## CI\/Gate Authority Stop Condition\n\n[^\n]+/u,
-        "",
-      );
-    published(root, env, automaticTask);
-    writeFileSync(path.join(root, "harness", automaticPlanPath), oneSectionMissing);
-    const automaticArgs = [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        "must acquire its lease",
-        "--task",
-        automaticTaskId,
-        "--detach",
-      ],
-      placeholderLease = runMaybe(root, env, automaticArgs);
-    assert.equal(placeholderLease.status, 1);
-    assert.equal(placeholderLease.receipt.code, "plan_placeholder");
-    assert.deepEqual(placeholderLease.receipt.diagnostic, {
-      kind: "missing-sections",
-      documentPath: automaticPlanPath,
-      diskDiffers: true,
-      missingSections: [{ section: "CI/Gate Authority Stop Condition", reason: "empty" }],
-    });
-    writeFileSync(path.join(root, "harness", automaticPackagePath, "task_plan.md"), realizedPlan("Automatic lease"));
-    const automaticPlanSync = run(root, env, ["doc", "sync", "--submit", "--path", automaticPlanPath]);
-    const automaticLease = runMaybe(root, env, automaticArgs);
-    assert.equal(automaticLease.status, 0, JSON.stringify(automaticLease));
-    assert.equal(automaticLease.receipt.outcome, "running");
-    assert.equal(typeof automaticLease.receipt.dispatchId, "string");
-    context.diagnostic(
-      `revision-aware plan hint: ${JSON.stringify({
-        before: {
-          outcome: placeholderLease.receipt.outcome,
-          code: placeholderLease.receipt.code,
-          diagnostic: placeholderLease.receipt.diagnostic,
-        },
-        sync: { outcome: automaticPlanSync.outcome, summary: automaticPlanSync.summary },
-        after: {
-          outcome: automaticLease.receipt.outcome,
-          dispatchId: automaticLease.receipt.dispatchId,
-        },
-      })}`,
-    );
-    const resumed = run(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt",
-      "second turn",
-      "--resume",
-      "provider-cli-session",
-      "--no-stream",
-    ]);
-    assert.equal((resumed.result as Record<string, unknown>).text, "resumed:provider-cli-session:second turn");
-    assert.equal((resumed.session as Record<string, unknown>).providerSessionId, "provider-cli-session");
-    const streamed = spawnSync(
-      process.execPath,
-      [cli, "--root", root, "runtime", "run", "cli-worker", "--prompt", "stream prompt"],
-      { encoding: "utf8", env },
-    );
-    assert.equal(streamed.status, 0, streamed.stderr);
-    assert.equal(streamed.stdout.trim(), "final:stream prompt");
-    const wildcard = run(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--agent",
-      "any-worker",
-      "--prompt",
-      "wildcard prompt",
-      "--no-stream",
-    ]);
-    const wildcardText = String((wildcard.result as Record<string, unknown>).text);
-    assert.ok(
-      wildcardText.startsWith(
-        "final:# Agent Identity: Any Worker (any-worker)\n\nUse any compatible runtime.\n\n# Harness Execution Discipline",
-      ),
-      wildcardText,
-    );
-    assert.match(wildcardText, /# Worker Role/u);
-    assert.ok(wildcardText.endsWith("\n\n# Mission\n\nwildcard prompt"), wildcardText);
-    assert.ok(
-      wildcardText.indexOf("# Worker Role") < wildcardText.indexOf("prompt://review") &&
-        wildcardText.indexOf("prompt://review") < wildcardText.indexOf("# Standard Task") &&
-        wildcardText.indexOf("# Standard Task") < wildcardText.indexOf("# Mission"),
-      "the role prompt, declared prompts, and preset must reach the real CLI dispatch in declaration order",
-    );
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const concurrentProjectionReads = await Promise.all([
-        ...Array.from({ length: 6 }, (_, index) =>
-          runAsync(root, env, [
-            "fact",
-            "record",
-            taskId,
-            "--statement",
-            `Concurrent projection observation ${String(index + 1)}`,
-            "--source",
-            `test:runtime-cli-concurrency:${String(index + 1)}`,
-          ]),
-        ),
-        runAsync(root, env, [
-          "runtime",
-          "run",
-          "cli-worker",
-          "--prompt",
-          `progress-middle:${taskId}`,
-          "--task",
-          taskId,
-          "--no-stream",
-        ]),
-      ]),
-      factWrites = concurrentProjectionReads.slice(0, 6),
-      progressResult = concurrentProjectionReads[6]!;
-    assert.equal(
-      factWrites.every((result) => result.status === 0 && result.receipt.outcome === "applied"),
-      true,
-      JSON.stringify(factWrites),
-    );
-    assert.equal(progressResult.status, 0, `${progressResult.stderr}\n${JSON.stringify(progressResult.receipt)}`);
-    context.diagnostic(
-      `projection readiness concurrency: ${JSON.stringify({ factWrites: factWrites.length, runtime: progressResult.receipt.outcome })}`,
-    );
-    const progressRun = progressResult.receipt,
-      progressDispatchId = String((progressRun.spawn as Record<string, unknown>).dispatchId),
-      shownProgress = JSON.parse(String(run(root, env, ["task", "show", taskId]).evidence)) as {
-        progress: Array<{
-          actor: { executor: unknown };
-          payload: { text: string; evidence: unknown[]; runtimeSessionId?: string };
-        }>;
-      },
-      runtimeExecutor = { kind: "agent", id: `runtime-session:${progressRun.runtimeSessionId}` };
-    assert.deepEqual(
-      shownProgress.progress.map((event) => event.payload.text),
-      ["Provider checkpoint one.", "Provider checkpoint two."],
-    );
-    assert.equal(
-      shownProgress.progress.every((event) => event.payload.evidence.length === 1),
-      true,
-    );
-    assert.deepEqual(
-      shownProgress.progress.map((event) => event.actor.executor),
-      [runtimeExecutor, runtimeExecutor],
-    );
-    assert.deepEqual(
-      shownProgress.progress.map((event) => event.payload.runtimeSessionId),
-      [progressRun.runtimeSessionId, progressRun.runtimeSessionId],
-    );
-    const unrelatedProgress = runMaybe(root, { ...env, HARNESS_ACTOR: "agent:unrelated-runtime" }, [
-      "task",
-      "progress",
-      "append",
-      taskId,
-      "--text",
-      "Unrelated checkpoint.",
-      "--evidence",
-      "test:reports/runtime-progress.txt:unrelated",
-    ]);
-    assert.equal(unrelatedProgress.status, 1);
-    assert.equal(unrelatedProgress.receipt.code, "progress_lease_required");
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const assembledPromptPrefix =
-        "# Agent Identity: Terra (terra)\n\nReview precisely.\n\n# Harness Execution Discipline",
-      bound = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--agent",
-        "fable",
-        "--to",
-        "terra",
-        "--prompt",
-        "bound prompt",
-        "--task",
-        taskId,
-        "--no-stream",
-      ]),
-      boundSpawn = bound.spawn as Record<string, unknown>,
-      boundDispatchId = String(boundSpawn.dispatchId),
-      boundDispatch = JSON.parse(
-        await readPublishedDispatch(path.join(artifactRoot, "dispatches", `${boundDispatchId}.json`)),
-      ) as Record<string, unknown>;
-    const assembledPrompt = readFileSync(path.join(artifactRoot, "missions", `${boundDispatchId}.md`), "utf8");
-    assert.ok(assembledPrompt.startsWith(assembledPromptPrefix), assembledPrompt);
-    assert.match(assembledPrompt, /# Worker Role/u);
-    assertTaskMissionPrompt(assembledPrompt, {
-      repoId: "runtime-cli",
-      taskId,
-      canonicalRoot: realpathSync(root),
-      workerRoot: realpathSync(root),
-      taskPackageRoot: path.join(realpathSync(root), "harness", packagePath),
-      daemonUserRoot: userRoot,
-      daemonId,
-      runtimeSessionId: String(bound.runtimeSessionId),
-      mission: "bound prompt",
-    });
-    assert.ok(
-      assembledPrompt.indexOf("# Worker Role") < assembledPrompt.indexOf("prompt://review") &&
-        assembledPrompt.indexOf("prompt://review") < assembledPrompt.indexOf("# Standard Task") &&
-        assembledPrompt.indexOf("# Standard Task") < assembledPrompt.indexOf("# Mission"),
-      "a squad-delegated worker must keep its role, prompts, and preset through the real CLI",
-    );
-    const sameNamedReportDispatches = [progressDispatchId, boundDispatchId],
-      sameNamedReportPaths = sameNamedReportDispatches.map((dispatchId) =>
-        path.join(artifactRoot, "reports", `${dispatchId}.md`),
-      ),
-      sameNamedReportArchives = sameNamedReportDispatches.map(
-        (dispatchId) =>
-          JSON.parse(readFileSync(path.join(artifactRoot, "dispatches", `${dispatchId}.json`), "utf8")) as Record<
-            string,
-            unknown
-          >,
-      );
-    assert.notEqual(progressDispatchId, boundDispatchId, "independent runtime reports must not share an artifact key");
-    assert.equal(sameNamedReportPaths.every(existsSync), true, "each runtime report must remain materialized");
-    assert.equal(
-      sameNamedReportArchives.every(
-        (archive, index) =>
-          archive.dispatchId === sameNamedReportDispatches[index] &&
-          archive.taskId === taskId &&
-          archive.executionId === executionId,
-      ),
-      true,
-      "each dispatch-scoped report must retain its own Task execution owner",
-    );
-    assert.notEqual(
-      readFileSync(sameNamedReportPaths[0]!, "utf8"),
-      readFileSync(sameNamedReportPaths[1]!, "utf8"),
-      "separate report bodies must not overwrite one another",
-    );
-    assert.equal(
-      readFileSync(path.join(artifactRoot, "reports", `${boundDispatchId}.md`), "utf8"),
-      `final:${assembledPrompt}`,
-    );
-    assert.deepEqual(boundDispatch, {
-      schema: "runtime-dispatch/v1",
-      dispatchId: boundDispatchId,
-      taskId,
-      executionId,
-      agentId: "terra",
-      agentName: "Terra",
-      delegatedByAgentId: "fable",
-      delegatedByAgentName: "Fable",
-      squadId: "core-squad",
-      instanceId: "cli-worker",
-      model: "runtime-test-model",
-      reasoningEffort: null,
-      fast: false,
-      cwd: realpathSync(root),
-      missionRef: `${packagePath}/artifacts/missions/${boundDispatchId}.md`,
-      runtimeSessionId: bound.runtimeSessionId,
-      providerSessionId: "provider-cli-session",
-      startedAt: boundDispatch.startedAt,
-      endedAt: boundDispatch.endedAt,
-      outcome: "succeeded",
-      exitCode: 0,
-      resultRef: (bound.result as Record<string, unknown>).ref,
-      eventStreamRef: `file:.harness/runtime/dispatches/${boundDispatchId}.jsonl`,
-      attemptGroupId: boundDispatchId,
-      attemptIndex: 0,
-      provider: { instance: "cli-worker", model: "runtime-test-model" },
-      classification: "worker_stop",
-      reason: "Worker completed the attempt successfully.",
-    });
-    assert.match(String(boundDispatch.startedAt), /^\d{4}-\d{2}-\d{2}T/u);
-    assert.match(String(boundDispatch.endedAt), /^\d{4}-\d{2}-\d{2}T/u);
-    assert.doesNotMatch(JSON.stringify(boundDispatch), /(?:api.?key|credential|environment|token)/iu);
-    assert.equal(existsSync(path.join(root, ".harness", "runtime", "dispatches", `${boundDispatchId}.jsonl`)), true);
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const dispatchRow = (
-      run(root, env, ["task", "dispatches", taskId]).dispatches as Array<Record<string, unknown>>
-    ).find((row) => row.dispatchId === boundDispatchId);
-    assert.deepEqual(
-      dispatchRow && {
-        agentId: dispatchRow.agentId,
-        delegatedByAgentId: dispatchRow.delegatedByAgentId,
-        squadId: dispatchRow.squadId,
-      },
-      { agentId: "terra", delegatedByAgentId: "fable", squadId: "core-squad" },
-    );
-    const outsider = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--agent",
-      "fable",
-      "--to",
-      "outsider",
-      "--prompt",
-      "reject outsider",
-      "--task",
-      taskId,
-      "--no-stream",
-    ]);
-    assert.equal(outsider.status, 1);
-    assert.equal(outsider.receipt.code, "squad_member_not_found");
-    const mismatch = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--agent",
-      "fable",
-      "--to",
-      "opencode-worker",
-      "--prompt",
-      "reject mismatch",
-      "--task",
-      taskId,
-      "--no-stream",
-    ]);
-    assert.equal(mismatch.status, 1);
-    assert.equal(mismatch.receipt.code, "agent_runtime_type_mismatch");
-    const existingMissionPath = `${packagePath}/artifacts/missions/existing-mission.md`;
-    const pathLikeMission = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--mission",
-      existingMissionPath,
-      "--task",
-      taskId,
-      "--no-stream",
-    ]);
-    assert.equal(pathLikeMission.status, 1);
-    assert.equal(pathLikeMission.receipt.code, "invalid_runtime_mission");
-    assert.deepEqual(pathLikeMission.receipt.diagnostic, {
-      kind: "validation",
-      entity: "runtime mission",
-      field: "mission",
-      actual: "path-like value",
-      expectation:
-        "Expected a bare mission id; the daemon resolves harness/<task-package>/artifacts/missions/<name>.md " +
-        "and did not look up this file. Retry ha runtime run <runtime-instance> --task <task-id> --mission <name>",
-    });
-    context.diagnostic(`invalid_runtime_mission receipt=${JSON.stringify(pathLikeMission.receipt)}`);
-    // The unavailable rejection is only deterministic while the file is absent: once it exists on
-    // disk, the WAL materializer's authored-candidate settlement may auto-submit it after any
-    // flush, racing both this rejection and a doc status that expects "eligible".
-    const unsyncedMission = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--mission",
-      "existing-mission",
-      "--task",
-      taskId,
-      "--no-stream",
-    ]);
-    assert.equal(unsyncedMission.status, 1);
-    assert.equal(unsyncedMission.receipt.code, "runtime_mission_unavailable");
-    mkdirSync(path.join(artifactRoot, "missions"), { recursive: true });
-    writeFileSync(path.join(root, "harness", existingMissionPath), "existing mission");
-    // Manual sync stays the tested path; a concurrent authored-candidate settlement may have won
-    // the write, which the receipt reports as no_changes — both outcomes leave the doc clean.
-    const missionSync = run(root, env, ["doc", "sync", "--submit", "--path", existingMissionPath]);
-    assert.match(String(run(root, env, ["doc", "status", "--path", existingMissionPath]).evidence), /"state":"clean"/u);
-    const retiredPromptFile = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt-file",
-      path.join("harness", packagePath, "artifacts", "missions", "existing-mission.md"),
-      "--task",
-      taskId,
-    ]);
-    assert.equal(retiredPromptFile.status, 2);
-    assert.equal(retiredPromptFile.receipt.code, "unknown_field");
-    const reused = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--mission",
-        "existing-mission",
-        "--task",
-        taskId,
-        "--no-stream",
-      ]),
-      reusedDispatchId = String((reused.spawn as Record<string, unknown>).dispatchId),
-      reusedDispatch = JSON.parse(
-        await readPublishedDispatch(path.join(artifactRoot, "dispatches", `${reusedDispatchId}.json`)),
-      ) as Record<string, unknown>;
-    const reusedMission = readFileSync(path.join(artifactRoot, "missions", `${reusedDispatchId}.md`), "utf8"),
-      reusedReport = readFileSync(path.join(artifactRoot, "reports", `${reusedDispatchId}.md`), "utf8");
-    context.diagnostic(
-      `hand-written mission route: ${JSON.stringify({
-        before: {
-          outcome: unsyncedMission.receipt.outcome,
-          code: unsyncedMission.receipt.code,
-          diagnostic: unsyncedMission.receipt.diagnostic,
-        },
-        sync: { outcome: missionSync.outcome, summary: missionSync.summary },
-        after: { outcome: reused.outcome, dispatchId: reusedDispatchId },
-      })}`,
-    );
-    assert.equal(reusedDispatch.missionRef, `${packagePath}/artifacts/missions/${reusedDispatchId}.md`);
-    assert.deepEqual(
-      readdirSync(path.join(artifactRoot, "missions")).sort(),
-      ["existing-mission.md", `${progressDispatchId}.md`, `${boundDispatchId}.md`, `${reusedDispatchId}.md`].sort(),
-    );
-    assert.equal(reusedReport, `final:${reusedMission}`);
-    assertTaskMissionPrompt(reusedMission, {
-      repoId: "runtime-cli",
-      taskId,
-      canonicalRoot: realpathSync(root),
-      workerRoot: realpathSync(root),
-      taskPackageRoot: path.join(realpathSync(root), "harness", packagePath),
-      daemonUserRoot: userRoot,
-      daemonId,
-      runtimeSessionId: String(reused.runtimeSessionId),
-      mission:
-        `Your task package is ${path.join(realpathSync(root), "harness", packagePath)}.\n` +
-        "Read task_plan.md in that package and complete the task.\n\n" +
-        "# Mission: existing-mission\n\nexisting mission",
-    });
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const taskPackage = path.join(realpathSync(root), "harness", packagePath),
-      derivedMission = `Your task package is ${taskPackage}.\nRead task_plan.md in that package and complete the task.`,
-      derived = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--agent",
-        "terra",
-        "--task",
-        taskId,
-        "--cwd",
-        ".",
-        "--no-stream",
-      ]),
-      derivedText = String((derived.result as Record<string, unknown>).text);
-    assert.match(
-      derivedText,
-      /^final:# Agent Identity: Terra \(terra\).*# Harness Execution Discipline.*# Worker Role/su,
-    );
-    assertTaskMissionPrompt(derivedText.slice("final:".length), {
-      repoId: "runtime-cli",
-      taskId,
-      canonicalRoot: realpathSync(root),
-      workerRoot: realpathSync(root),
-      taskPackageRoot: taskPackage,
-      daemonUserRoot: userRoot,
-      daemonId,
-      runtimeSessionId: String(derived.runtimeSessionId),
-      mission: derivedMission,
-    });
-    assert.equal(derivedText.includes("Review precisely."), true);
-    const listed = run(root, env, ["runtime", "status"]),
-      sessions = listed.sessions as Array<Record<string, unknown>>;
-    assert.equal(sessions.length, 9);
-    assert.match(streamed.stderr, /\[message\] live:stream prompt/u, JSON.stringify(sessions));
-    const detail = run(root, env, ["runtime", "status", String(resumed.runtimeSessionId)]);
-    assert.equal((detail.result as Record<string, unknown>).text, "resumed:provider-cli-session:second turn");
-    const waited = run(root, env, ["runtime", "status", String(resumed.runtimeSessionId), "--wait", "--no-stream"]);
-    assert.equal(waited.command, "runtime-status");
-    assert.equal(waited.summary, "resumed:provider-cli-session:second turn");
-    const detachedProcess = runMaybe(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        "stream prompt",
-        "--detach",
-      ]),
-      detachedReceipt = detachedProcess.receipt,
-      detachedRuntimeSessionId = String(detachedReceipt.runtimeSessionId);
-    assert.equal(detachedProcess.status, 0, detachedProcess.stderr);
-    assert.equal(detachedReceipt.nextAction, `ha runtime status ${detachedRuntimeSessionId} --wait`);
-    assert.match(
-      String(detachedReceipt.summary),
-      new RegExp(`next: ha runtime status ${detachedRuntimeSessionId} --wait$`, "u"),
-    );
-    const retrievalProcess = runMaybe(root, env, [
-        "runtime",
-        "status",
-        detachedRuntimeSessionId,
-        "--wait",
-        "--no-stream",
-      ]),
-      retrievedText = String((retrievalProcess.receipt.result as Record<string, unknown>).text);
-    assert.equal(retrievalProcess.status, 0, retrievalProcess.stderr);
-    assert.equal(retrievalProcess.receipt.outcome, "succeeded");
-    assert.equal(retrievedText, "final:stream prompt");
-    assert.equal(Number.isInteger(detachedProcess.pid) && Number.isInteger(retrievalProcess.pid), true);
-    assert.notEqual(
-      detachedProcess.pid,
-      retrievalProcess.pid,
-      "detach and status --wait must execute in different CLI processes",
-    );
-    const readerReuse = await eventuallyRuntimeReaderReuse(userRoot, daemonId);
-    context.diagnostic(
-      `detach cross-process retrieval: ${JSON.stringify({ launcherPid: detachedProcess.pid, retrievalPid: retrievalProcess.pid, runtimeSessionId: detachedRuntimeSessionId, outcome: retrievalProcess.receipt.outcome, resultText: retrievedText, readerReuse })}`,
-    );
-    const restartDetached = run(root, env, ["runtime", "run", "cli-worker", "--prompt", "hold", "--detach"]),
-      restartSessionId = String(restartDetached.runtimeSessionId),
-      restartDispatchId = String(restartDetached.dispatchId),
-      liveBeforeRestart = await eventuallyRuntimeStatus(root, env, restartSessionId, "live"),
-      workerPid = Number(
-        readDispatchRecords(root, restartDispatchId).find((record) => record.kind === "process_started")?.pid,
-      ),
-      daemonBeforeRestart = run(root, env, ["daemon", "status"]),
-      stoppedForRestart = run(root, env, ["daemon", "stop", "--force"]);
-    assert.equal(liveBeforeRestart.session.liveness, "live");
-    assert.equal(
-      processAlive(workerPid),
-      true,
-      `runtime worker ${String(workerPid)} must be live before daemon restart`,
-    );
-    assert.equal(stoppedForRestart.forced, true, JSON.stringify(stoppedForRestart));
-    assert.equal(processAlive(workerPid), true, `runtime worker ${String(workerPid)} must survive daemon stop --force`);
-    assert.equal(run(root, env, ["daemon", "start", "--service"]).ok, true);
-    const liveAfterRestart = await eventuallyRuntimeStatus(root, env, restartSessionId, "live"),
-      daemonAfterRestart = run(root, env, ["daemon", "status"]);
-    assert.notEqual(daemonAfterRestart.pid, daemonBeforeRestart.pid, "runtime status must use a restarted daemon");
-    assert.equal(liveAfterRestart.session.liveness, "live");
-    assert.equal(liveAfterRestart.session.activity.outcome, null);
-    assert.equal(processAlive(workerPid), true, `runtime worker ${String(workerPid)} must agree with runtime status`);
-    assert.equal(run(root, env, ["runtime", "cancel", restartSessionId]).detail, "cancelled");
-    context.diagnostic(
-      `daemon restart liveness: ${JSON.stringify({
-        runtimeSessionId: restartSessionId,
-        workerPid,
-        daemonBefore: daemonBeforeRestart.pid,
-        daemonAfter: daemonAfterRestart.pid,
-        before: liveBeforeRestart.session.liveness,
-        after: liveAfterRestart.session.liveness,
-      })}`,
-    );
-    const invalidOnExit = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt",
-      "invalid callback",
-      "--on-exit",
-      nonzeroNotifier,
-    ]);
-    assert.equal(invalidOnExit.status, 2);
-    assert.equal(invalidOnExit.receipt.code, "invalid_field");
-    assert.equal((invalidOnExit.receipt.diagnostic as Record<string, unknown>).kind, "validation");
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const controlNotification = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        "notification control",
-        "--task",
-        taskId,
-        "--detach",
-      ]),
-      controlNotificationWait = run(root, env, [
-        "runtime",
-        "status",
-        String(controlNotification.runtimeSessionId),
-        "--wait",
-        "--no-stream",
-      ]),
-      controlInvariant = await runtimeInvariantEvidence(
-        root,
-        artifactRoot,
-        controlNotification,
-        controlNotificationWait,
-      );
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const missingNotifier = path.join(parent, "missing-notifier"),
-      missingNotification = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        "notification missing",
-        "--task",
-        taskId,
-        "--detach",
-        "--on-exit",
-        missingNotifier,
-      ]),
-      missingNotificationWait = run(root, env, [
-        "runtime",
-        "status",
-        String(missingNotification.runtimeSessionId),
-        "--wait",
-        "--no-stream",
-      ]),
-      missingTrace = await eventuallyNotification(root, String(missingNotification.dispatchId)),
-      missingInvariant = await runtimeInvariantEvidence(
-        root,
-        artifactRoot,
-        missingNotification,
-        missingNotificationWait,
-      );
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const nonzeroNotification = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        "notification nonzero",
-        "--task",
-        taskId,
-        "--detach",
-        "--on-exit",
-        nonzeroNotifier,
-      ]),
-      nonzeroNotificationWait = run(root, env, [
-        "runtime",
-        "status",
-        String(nonzeroNotification.runtimeSessionId),
-        "--wait",
-        "--no-stream",
-      ]),
-      nonzeroNotificationTrace = await eventuallyNotification(root, String(nonzeroNotification.dispatchId)),
-      nonzeroInvariant = await runtimeInvariantEvidence(
-        root,
-        artifactRoot,
-        nonzeroNotification,
-        nonzeroNotificationWait,
-      ),
-      callbackObservation = JSON.parse(readFileSync(nonzeroTrace, "utf8")) as Record<string, unknown>,
-      callbackEnvironment = callbackObservation.environment as Record<string, unknown>,
-      callbackPayload = callbackObservation.payload as Record<string, unknown>;
-    assert.deepEqual(missingInvariant, controlInvariant);
-    assert.deepEqual(nonzeroInvariant, controlInvariant);
-    assert.deepEqual(
-      missingTrace.finished && {
-        started: missingTrace.finished.started,
-        exitCode: missingTrace.finished.exitCode,
-        timedOut: missingTrace.finished.timedOut,
-        errorCode: missingTrace.finished.errorCode,
-      },
-      { started: false, exitCode: null, timedOut: false, errorCode: "ENOENT" },
-    );
-    assert.deepEqual(
-      nonzeroNotificationTrace.finished && {
-        started: nonzeroNotificationTrace.finished.started,
-        exitCode: nonzeroNotificationTrace.finished.exitCode,
-        timedOut: nonzeroNotificationTrace.finished.timedOut,
-      },
-      { started: true, exitCode: 23, timedOut: false },
-    );
-    assert.equal(callbackObservation.cwd, realpathSync(root));
-    assert.equal(callbackEnvironment.OPENAI_API_KEY, undefined);
-    assert.equal(callbackEnvironment.HARNESS_NOTIFY_TEST_SECRET, undefined);
-    assert.deepEqual(callbackPayload, {
-      schema: "runtime-session-exited/v1",
-      runtimeSessionId: nonzeroNotification.runtimeSessionId,
-      outcome: "succeeded",
-      exitCode: 0,
-    });
-    assert.doesNotMatch(JSON.stringify(callbackPayload), /notification nonzero|final:|credential|token|api.?key/iu);
-    const missingArchive = JSON.parse(
-        await readPublishedDispatch(
-          path.join(artifactRoot, "dispatches", `${String(missingNotification.dispatchId)}.json`),
-        ),
-      ) as Record<string, unknown>,
-      nonzeroArchive = JSON.parse(
-        await readPublishedDispatch(
-          path.join(artifactRoot, "dispatches", `${String(nonzeroNotification.dispatchId)}.json`),
-        ),
-      ) as Record<string, unknown>;
-    assert.equal(missingArchive.onExitCommand, missingNotifier);
-    assert.equal(nonzeroArchive.onExitCommand, nonzeroNotifier);
-    context.diagnostic(
-      `failed notification trace: ${JSON.stringify({ missing: missingTrace.finished, nonzero: nonzeroNotificationTrace.finished, invariant: nonzeroInvariant })}`,
-    );
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const queueNotification = run(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt",
-      "notification queue hold",
-      "--task",
-      taskId,
-      "--detach",
-      "--on-exit",
-      holdNotifier,
-    ]);
-    run(root, env, ["runtime", "status", String(queueNotification.runtimeSessionId), "--wait", "--no-stream"]);
-    await eventuallyFile(notificationStarted);
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const revisionBefore = makeTaskEventReader({ repoId: "runtime-cli", rootDir: root }).readHead()?.revision ?? 0,
-      queueWrite = run(root, env, [
-        "task",
-        "progress",
-        "append",
-        taskId,
-        "--text",
-        "Notification queue remained writable.",
-        "--evidence",
-        "test:reports/notification-queue.txt:write progressed while callback was active",
-      ]),
-      revisionAfter = makeTaskEventReader({ repoId: "runtime-cli", rootDir: root }).readHead()?.revision ?? 0;
-    assert.ok(revisionAfter > revisionBefore, `${revisionBefore} -> ${revisionAfter}`);
-    assert.equal(
-      existsSync(notificationFinished),
-      false,
-      "the notification must still be executing when the repo write commits",
-    );
-    writeFileSync(notificationRelease, "release\n");
-    const queueNotificationTrace = await eventuallyNotification(root, String(queueNotification.dispatchId));
-    assert.equal(queueNotificationTrace.started?.phase, "started");
-    assert.equal(queueNotificationTrace.finished?.timedOut, false);
-    assert.equal(existsSync(notificationFinished), true);
-    context.diagnostic(
-      `notification queue concurrency: ${JSON.stringify({ revisionBefore, revisionAfter, writeReceiptRevision: queueWrite.revision, notifierStillRunningAtCommit: true })}`,
-    );
-    const submittedTaskId = `${taskId}-submitted`,
-      submittedExecutionId = `${executionId}-submitted`,
-      submittedTask = run(root, env, [
-        "task",
-        "create",
-        "--id",
-        submittedTaskId,
-        "--admin",
-        "--title",
-        "Submitted runtime archive",
-      ]),
-      submittedPackagePath = String(submittedTask.packagePath),
-      submittedPlanPath = `${submittedPackagePath}/task_plan.md`;
-    published(root, env, submittedTask);
-    writeFileSync(path.join(root, "harness", submittedPlanPath), realizedPlan("Submitted runtime archive"));
-    run(root, env, ["doc", "sync", "--submit", "--path", submittedPlanPath]);
-    run(root, env, ["task", "start", submittedTaskId, "--execution-id", submittedExecutionId]);
-    // This standard task submits a public code cut bound to the runtime cwd.
-    const deliveryGit = (args: string[]) => {
-      const result = spawnSync("git", args, { cwd: root, env, encoding: "utf8" });
-      assert.equal(result.status, 0, result.stderr);
-      return result.stdout.trim();
-    };
-    deliveryGit(["init", "-b", "main"]);
-    deliveryGit(["config", "user.name", "Harness Test"]);
-    deliveryGit(["config", "user.email", "harness@example.test"]);
-    writeFileSync(path.join(root, "README.md"), "Runtime archive baseline.\n");
-    deliveryGit(["add", "README.md"]);
-    deliveryGit(["commit", "-m", "test: seed runtime delivery baseline"]);
-    deliveryGit(["update-ref", "refs/remotes/origin/main", "HEAD"]);
-    writeFileSync(path.join(root, "README.md"), "Runtime archive delivery.\n");
-    deliveryGit(["add", "README.md"]);
-    deliveryGit(["commit", "-m", "test: record runtime delivery"]);
-    const submittedCut = { commitSha: deliveryGit(["rev-parse", "HEAD"]) };
-    const submittedRuntime = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        `submit-before-exit:${submittedTaskId}:${submittedCut.commitSha}`,
-        "--task",
-        submittedTaskId,
-        "--no-stream",
-      ]),
-      submittedDispatchId = String((submittedRuntime.spawn as Record<string, unknown>).dispatchId),
-      submittedReportPath = `${submittedPackagePath}/artifacts/reports/${submittedDispatchId}.md`,
-      submittedDispatch = (
-        run(root, env, ["task", "dispatches", submittedTaskId]).dispatches as Array<Record<string, unknown>>
-      ).find((row) => row.dispatchId === submittedDispatchId);
-    await eventuallyFile(path.join(root, "harness", submittedReportPath));
-    assert.equal(existsSync(path.join(root, "harness", submittedReportPath)), true);
-    assert.equal(submittedDispatch?.reportPath, submittedReportPath);
-    const readOnly = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--permission-mode",
-      "read-only",
-      "--prompt",
-      "read-only",
-      "--no-stream",
-    ]);
-    assert.equal(readOnly.status, 0, `${readOnly.stderr}\n${JSON.stringify(readOnly.receipt)}`);
-    assert.equal(readOnly.receipt.outcome, "succeeded");
-    const readOnlySpawn = readOnly.receipt.spawn as Record<string, unknown>;
-    assert.deepEqual(
-      { ledgerAccess: readOnlySpawn.ledgerAccess, reportDelivery: readOnlySpawn.reportDelivery },
-      { ledgerAccess: "unavailable", reportDelivery: "stdout" },
-    );
-    assert.match(
-      String((readOnly.receipt.result as Record<string, unknown>).text),
-      /Read-only Dispatch Contract[\s\S]*daemon-ledger commands are unavailable[\s\S]*final stdout/u,
-    );
-    const noAction = runMaybe(root, env, ["runtime", "run", "cli-worker", "--prompt", "no-action", "--no-stream"]);
-    assert.equal(noAction.status, 0, `${noAction.stderr}\n${JSON.stringify(noAction.receipt)}`);
-    assert.equal(noAction.receipt.outcome, "succeeded");
-    writeFileSync(
-      path.join(root, "batch-unknown-declaration.json"),
-      JSON.stringify({
-        schema: "runtime-batch/v1",
-        maxConcurrency: 1,
-        dispatches: [{ instance: "cli-worker", prompt: "unused" }],
-        permissionMode: "read-only",
-      }),
-    );
-    const unknownDeclaration = runMaybe(root, env, ["runtime", "batch", "batch-unknown-declaration.json"]);
-    assert.equal(unknownDeclaration.status, 1);
-    assert.equal(unknownDeclaration.receipt.code, "batch_file_invalid");
-    assert.equal((unknownDeclaration.receipt.diagnostic as Record<string, unknown>).kind, "validation");
-    writeFileSync(
-      path.join(root, "batch-unknown-dispatch.json"),
-      JSON.stringify({
-        schema: "runtime-batch/v1",
-        maxConcurrency: 1,
-        dispatches: [{ instance: "cli-worker", prompt: "unused", permissionMode: "read-only" }],
-      }),
-    );
-    const unknownDispatch = runMaybe(root, env, ["runtime", "batch", "batch-unknown-dispatch.json"]);
-    assert.equal(unknownDispatch.status, 1);
-    assert.equal(unknownDispatch.receipt.code, "batch_file_invalid");
-    assert.equal((unknownDispatch.receipt.diagnostic as Record<string, unknown>).kind, "validation");
-    const unknownSpawn = await runCommandThroughDaemon(
-      {
-        rootDir: safePath(root),
-        repoId: "runtime-cli",
-        json: true,
-        method: "repo.agentRuntime.spawn",
-        action: {
-          kind: "runtime-run",
-          runtimeInstanceId: "cli-worker",
-          cwd: { scope: "repo-root" },
-          prompt: "unused",
-          taskId: null,
-          idempotencyKey: "unknown-wire-field",
-          permission_mode: "read-only",
-        } as never,
-      },
-      undefined,
-      { env },
-    );
-    assert.equal(unknownSpawn.code, "unknown_field");
-    const unknownSpawnDiagnostic = unknownSpawn.diagnostic as Record<string, unknown>;
-    assert.deepEqual(
-      {
-        kind: unknownSpawnDiagnostic.kind,
-        entity: unknownSpawnDiagnostic.entity,
-        field: unknownSpawnDiagnostic.field,
-        actual: unknownSpawnDiagnostic.actual,
-      },
-      {
-        kind: "validation",
-        entity: "repo.agentRuntime.spawn",
-        field: "permission_mode",
-        actual: "unknown",
-      },
-    );
-    assert.match(String(unknownSpawnDiagnostic.expectation), /Allowed fields:.*agentId.*permissionMode/u);
-    const unknownCancel = await runCommandThroughDaemon(
-      {
-        rootDir: safePath(root),
-        repoId: "runtime-cli",
-        json: true,
-        method: "repo.agentRuntime.cancel",
-        action: { kind: "runtime-cancel", runtimeSessionId: "missing", force: true } as never,
-      },
-      undefined,
-      { env },
-    );
-    assert.equal(unknownCancel.code, "unknown_field");
-    assert.deepEqual(
-      {
-        kind: (unknownCancel.diagnostic as Record<string, unknown>).kind,
-        field: (unknownCancel.diagnostic as Record<string, unknown>).field,
-      },
-      { kind: "validation", field: "force" },
-    );
-    const unknownFact = await runCommandThroughDaemon(
-      {
-        rootDir: safePath(root),
-        repoId: "runtime-cli",
-        json: true,
-        method: "repo.task.read",
-        action: { kind: "fact-search", taskId, permissionMode: "read-only" } as never,
-      },
-      undefined,
-      { env: { ...env, HARNESS_ACTOR: "" } },
-    );
-    assert.equal(unknownFact.code, "invalid_command");
-    writeFileSync(
-      path.join(root, "batch.json"),
-      JSON.stringify(
-        {
-          schema: "runtime-batch/v1",
-          maxConcurrency: 2,
-          dispatches: [
-            { instance: "cli-worker", agent: "fable", to: "missing-agent", prompt: "must reject" },
-            {
-              instance: "cli-worker",
-              agent: "fable",
-              to: "terra",
-              model: "runtime-test-model",
-              effort: "high",
-              fast: true,
-              prompt: "batch hold one",
-              task: taskId,
-            },
-            {
-              instance: "cli-worker",
-              agent: "fable",
-              to: "terra",
-              effort: "high",
-              fast: true,
-              prompt: "batch hold two",
-              cwd: ".",
-              task: taskId,
-            },
-            {
-              instance: "cli-worker",
-              agent: "fable",
-              to: "terra",
-              effort: "high",
-              fast: true,
-              mission: "existing-mission",
-              task: taskId,
-            },
-            {
-              instance: "cli-worker",
-              agent: "fable",
-              to: "terra",
-              effort: "high",
-              fast: true,
-              prompt: "batch hold three",
-              task: taskId,
-            },
-          ],
-        },
-        null,
-        2,
-      ),
-    );
-    const batch = runMaybe(root, env, ["runtime", "batch", "batch.json"]);
-    assert.equal(batch.status, 1, `${batch.stderr}\n${JSON.stringify(batch.receipt)}`);
-    assert.equal(batch.receipt.command, "runtime-batch");
-    assert.equal(batch.receipt.outcome, "partial_failure");
-    const batchRows = batch.receipt.dispatches as Array<Record<string, unknown>>;
-    assert.equal(batchRows.length, 5);
-    const taskBatchRows = batchRows.slice(1),
-      successfulBatchRows = taskBatchRows.filter((row) => row.status === "succeeded"),
-      rejectedTaskBatchRows = taskBatchRows.filter((row) => row.status === "rejected");
-    assert.ok(successfulBatchRows.length >= 1, JSON.stringify(batch.receipt));
-    assert.equal(
-      successfulBatchRows.length + rejectedTaskBatchRows.length,
-      taskBatchRows.length,
-      JSON.stringify(batch.receipt),
-    );
-    assert.equal(
-      rejectedTaskBatchRows.every((row) => row.code === "runtime_task_lease_required"),
-      true,
-      JSON.stringify(batch.receipt),
-    );
-    assert.equal(batchRows[0]?.code, "squad_member_not_found");
-    assert.equal(
-      successfulBatchRows.every(
-        (row) => typeof row.dispatchId === "string" && typeof row.runtimeSessionId === "string",
-      ),
-      true,
-    );
-    const batchDispatchIds = new Set(successfulBatchRows.map((row) => String(row.dispatchId))),
-      batchDispatches = (
-        run(root, env, ["task", "dispatches", taskId]).dispatches as Array<Record<string, unknown>>
-      ).filter((row) => batchDispatchIds.has(String(row.dispatchId)));
-    assert.equal(batchDispatches.length, successfulBatchRows.length);
-    assert.equal(
-      batchDispatches.every(
-        (row) => row.agentId === "terra" && row.delegatedByAgentId === "fable" && row.squadId === "core-squad",
-      ),
-      true,
-      JSON.stringify(batchDispatches),
-    );
-    const batchArchives = successfulBatchRows.map((row) =>
-      path.join(artifactRoot, "dispatches", `${String(row.dispatchId)}.json`),
-    );
-    await Promise.all(batchArchives.map(eventuallyFile));
-    assert.deepEqual(
-      batchArchives.filter((archive) => !existsSync(archive)),
-      [],
-      "successful runtime batch dispatch archives must become visible",
-    );
-    const slowBatchArchive = batchArchives[0]!;
-    assert.equal(
-      (JSON.parse(readFileSync(slowBatchArchive, "utf8")) as Record<string, unknown>).reasoningEffort,
-      "high",
-    );
-    assert.equal((JSON.parse(readFileSync(slowBatchArchive, "utf8")) as Record<string, unknown>).fast, true);
-    const emptyFailure = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt",
-      "failure:empty",
-      "--no-stream",
-    ]);
-    assert.equal(emptyFailure.status, 1);
-    assert.equal(emptyFailure.receipt.code, "provider_exit");
-    assert.equal(emptyFailure.receipt.reason, "Provider exited with code 1 and produced no output.");
-    assert.equal(emptyFailure.receipt.summary, emptyFailure.receipt.reason);
-    const secret = "sk-runtime-secret-1234567890",
-      secretFailure = runMaybe(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        "failure:secret",
-        "--no-stream",
-      ]);
-    assert.equal(secretFailure.status, 1);
-    assert.match(String(secretFailure.receipt.reason), /Provider exited with code 1.*OPENAI_API_KEY=\[REDACTED\]/u);
-    assert.doesNotMatch(JSON.stringify(secretFailure.receipt), new RegExp(secret, "u"));
-    assert.doesNotMatch(
-      readFileSync(path.join(root, ".harness", "requests", "requests.jsonl"), "utf8"),
-      new RegExp(secret, "u"),
-    );
-    const structuredFailure = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt",
-      "failure:structured",
-      "--no-stream",
-    ]);
-    assert.equal(structuredFailure.status, 1);
-    assert.match(String(structuredFailure.receipt.reason), /structured provider failure/u);
-    assert.doesNotMatch(JSON.stringify(structuredFailure.receipt), new RegExp(secret, "u"));
-    const quotaFailure = runMaybe(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt",
-      "failure:429",
-      "--no-stream",
-    ]);
-    assert.equal(quotaFailure.status, 1);
-    assert.equal(quotaFailure.receipt.code, "quota_exhausted");
-    assert.match(
-      String(quotaFailure.receipt.reason),
-      /faultClass=quota_exhausted; resetAt=2026-09-06T05:06:07\.000Z;.*credit balance exhausted/u,
-    );
-    writeFileSync(
-      path.join(root, "failure-batch.json"),
-      JSON.stringify({
-        schema: "runtime-batch/v1",
-        maxConcurrency: 1,
-        dispatches: [{ instance: "cli-worker", prompt: "failure:empty" }],
-      }),
-    );
-    const failureBatch = runMaybe(root, env, ["runtime", "batch", "failure-batch.json"]);
-    assert.equal(failureBatch.status, 1);
-    const failureRow = (failureBatch.receipt.dispatches as Array<Record<string, unknown>>)[0]!;
-    assert.equal(failureRow.code, "provider_exit");
-    assert.equal(failureRow.reason, "Provider exited with code 1 and produced no output.");
-    const taskWait = runMaybe(root, env, ["runtime", "status", "--task", taskId, "--wait", "--no-stream"]);
-    assert.equal(taskWait.status, 0, `${taskWait.stderr}\n${JSON.stringify(taskWait.receipt)}`);
-    assert.equal(taskWait.receipt.outcome, "succeeded");
-    const waitedRow = (taskWait.receipt.dispatches as Array<Record<string, unknown>>).find(
-      (row) => row.dispatchId === progressDispatchId,
-    );
-    assert.equal(waitedRow?.exitCode, 0);
-    assert.equal(waitedRow?.dispatchPath, `${packagePath}/artifacts/dispatches/${progressDispatchId}.json`);
-    assert.equal(waitedRow?.reportPath, `${packagePath}/artifacts/reports/${progressDispatchId}.md`);
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const detachedFailure = run(root, env, [
-      "runtime",
-      "run",
-      "cli-worker",
-      "--prompt",
-      "failure:structured",
-      "--task",
-      taskId,
-      "--detach",
-    ]);
-    const failureWait = runMaybe(root, env, ["runtime", "status", "--task", taskId, "--wait", "--no-stream"]);
-    assert.equal(detachedFailure.outcome, "running");
-    assert.equal(failureWait.status, 1, `${failureWait.stderr}\n${JSON.stringify(failureWait.receipt)}`);
-    assert.equal(failureWait.receipt.outcome, "failed");
-    const failedRow = (failureWait.receipt.dispatches as Array<Record<string, unknown>>).find(
-      (row) => row.dispatchId === detachedFailure.dispatchId,
-    );
-    assert.equal(failedRow?.status, "failed");
-    assert.equal(failedRow?.exitCode, 1);
-    assert.equal(failedRow?.dispatchPath, `${packagePath}/artifacts/dispatches/${detachedFailure.dispatchId}.json`);
-    await eventuallyFile(path.join(artifactRoot, "reports", `${detachedFailure.dispatchId}.md`));
-    const publishedFailureRow = (
-      run(root, env, ["task", "dispatches", taskId]).dispatches as Array<Record<string, unknown>>
-    ).find((row) => row.dispatchId === detachedFailure.dispatchId);
-    assert.equal(publishedFailureRow?.reportPath, `${packagePath}/artifacts/reports/${detachedFailure.dispatchId}.md`);
-    const healthAfterRuntimeFailure = (run(root, env, ["daemon", "status"]).repos as Array<Record<string, unknown>>)[0]
-      ?.materialization as Record<string, unknown>;
-    assert.equal(healthAfterRuntimeFailure.state, "ok", JSON.stringify(healthAfterRuntimeFailure));
-    context.diagnostic(
-      `detach -> task wait failure: ${JSON.stringify({
-        detached: detachedFailure,
-        wait: failureWait.receipt,
-        dispatch: failedRow,
-        materialization: healthAfterRuntimeFailure,
-      })}`,
-    );
-    const events = existsSync(tracker) ? readFileSync(tracker, "utf8").trim().split("\n").filter(Boolean) : [],
-      trackedBatchDispatches = successfulBatchRows.filter((row) => row.index !== 3).length;
-    let active = 0,
-      peak = 0;
-    for (const event of events) {
-      active += event === "start" ? 1 : -1;
-      peak = Math.max(peak, active);
-    }
-    assert.equal(events.filter((event) => event === "start").length, trackedBatchDispatches);
-    assert.equal(events.filter((event) => event === "end").length, trackedBatchDispatches);
-    assert.ok(peak <= 1, `task-bound batch exceeded its single lease holder: ${events.join(",")}`);
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const detached = run(root, env, [
-        "runtime",
-        "run",
-        "cli-worker",
-        "--prompt",
-        "hold",
-        "--task",
-        taskId,
-        "--detach",
-        "--on-exit",
-        onceNotifier,
-      ]),
-      detachedDispatchId = String(detached.dispatchId),
-      detachedSessionId = String(detached.runtimeSessionId),
-      running = run(root, env, ["task", "dispatches", taskId]);
-    assert.equal(detached.outcome, "running");
-    assert.equal(
-      (running.dispatches as Array<Record<string, unknown>>).find((row) => row.dispatchId === detachedDispatchId)
-        ?.status,
-      "running",
-    );
-    assert.equal(run(root, env, ["runtime", "cancel", detachedSessionId]).detail, "cancelled");
-    assert.equal(run(root, env, ["runtime", "cancel", detachedSessionId]).detail, "already-exited");
-    const cancelled = await eventually(() => run(root, env, ["task", "dispatches", taskId])),
-      cancelledRow = (cancelled.dispatches as Array<Record<string, unknown>>).find(
-        (row) => row.dispatchId === detachedDispatchId,
-      )!;
-    assert.equal(cancelledRow.status, "cancelled");
-    assert.equal(
-      (
-        JSON.parse(
-          await readPublishedDispatch(path.join(artifactRoot, "dispatches", `${detachedDispatchId}.json`)),
-        ) as Record<string, unknown>
-      ).outcome,
-      "cancelled",
-    );
-    const cancelledReport = readFileSync(path.join(artifactRoot, "reports", `${detachedDispatchId}.md`), "utf8");
-    assertTaskMissionPrompt(cancelledReport.slice("live:".length), {
-      repoId: "runtime-cli",
-      taskId,
-      canonicalRoot: realpathSync(root),
-      workerRoot: realpathSync(root),
-      taskPackageRoot: path.join(realpathSync(root), "harness", packagePath),
-      daemonUserRoot: userRoot,
-      daemonId,
-      runtimeSessionId: detachedSessionId,
-      mission: "hold",
-    });
-    const cancelledWait = runMaybe(root, env, ["runtime", "status", detachedSessionId, "--wait", "--no-stream"]);
-    assert.equal(cancelledWait.status, 1);
-    assert.equal(cancelledWait.receipt.outcome, "cancelled");
-    const cancelledNotificationTrace = await eventuallyNotification(root, detachedDispatchId);
-    await eventuallyFile(notificationOnce);
-    const onceLines = readFileSync(notificationOnce, "utf8").trim().split(/\r?\n/u),
-      cancelledRecords = readDispatchRecords(root, detachedDispatchId).filter(
-        (record) => record.kind === "exit_notification",
-      );
-    assert.equal(onceLines.length, 1, JSON.stringify(onceLines));
-    assert.equal(cancelledRecords.filter((record) => record.phase === "started").length, 1);
-    assert.equal(cancelledRecords.filter((record) => record.phase === "finished").length, 1);
-    assert.equal(cancelledNotificationTrace.finished?.exitCode, 0);
-    context.diagnostic(
-      `notification at-most-once: ${JSON.stringify({ executions: onceLines.length, records: cancelledRecords })}`,
-    );
-    run(root, env, ["task", "start", taskId, "--execution-id", executionId]);
-    const streamRoot = path.join(root, ".harness", "runtime", "dispatches"),
-      streamsBeforeRejectedResume = readdirSync(streamRoot).sort(),
-      sessionsBeforeRejectedResume = (run(root, env, ["runtime", "status"]).sessions as Array<Record<string, unknown>>)
-        .length,
-      rejectedResume = runMaybe(root, env, [
-        "runtime",
-        "run",
-        "--resume-dispatch",
-        detachedDispatchId,
-        "--prompt",
-        "failure:empty",
-        "--detach",
-      ]);
-    assert.equal(rejectedResume.status, 1);
-    assert.equal(rejectedResume.receipt.code, "runtime_resume_failed");
-    assert.deepEqual(readdirSync(streamRoot).sort(), streamsBeforeRejectedResume);
-    assert.equal(
-      (run(root, env, ["runtime", "status"]).sessions as Array<Record<string, unknown>>).length,
-      sessionsBeforeRejectedResume,
-    );
-    const resumedDispatch = run(root, env, [
-        "runtime",
-        "run",
-        "--resume-dispatch",
-        detachedDispatchId,
-        "--prompt",
-        "follow up",
-        "--no-stream",
-      ]),
-      resumedDispatchId = String((resumedDispatch.spawn as Record<string, unknown>).dispatchId),
-      resumedText = String((resumedDispatch.result as Record<string, unknown>).text);
-    assert.ok(resumedText.startsWith("resumed:provider-cli-session:"), resumedText);
-    assertTaskMissionPrompt(resumedText.slice("resumed:provider-cli-session:".length), {
-      repoId: "runtime-cli",
-      taskId,
-      canonicalRoot: realpathSync(root),
-      workerRoot: realpathSync(root),
-      taskPackageRoot: path.join(realpathSync(root), "harness", packagePath),
-      daemonUserRoot: userRoot,
-      daemonId,
-      runtimeSessionId: String(resumedDispatch.runtimeSessionId),
-      mission: "follow up",
-    });
-    const resumedRow = (
-      run(root, env, ["task", "dispatches", taskId]).dispatches as Array<Record<string, unknown>>
-    ).find((row) => row.dispatchId === resumedDispatchId);
-    assert.equal(resumedRow?.status, "succeeded");
-    assert.equal(
-      (
-        JSON.parse(
-          await readPublishedDispatch(path.join(artifactRoot, "dispatches", `${resumedDispatchId}.json`)),
-        ) as Record<string, unknown>
-      ).providerSessionId,
-      "provider-cli-session",
-    );
-    // #1572: with HARNESS_ACTOR declared, the daemon request log is the server-side proof that executor
-    // attribution still arrives — inside the action for repo.task.run writes (task start) and at payload
-    // level for preset methods (task create) — while nothing is ever rejected for an undeclared executor.
-    const requests = readFileSync(path.join(root, ".harness", "requests", "requests.jsonl"), "utf8")
-      .trim()
-      .split("\n")
-      .map((line) => JSON.parse(line) as Record<string, unknown>);
-    assert.deepEqual(
-      requests.find((entry) => entry.method === "repo.task.run" && entry.command === "task-start")?.executor,
-      { kind: "agent", id: "runtime-cli-test" },
-    );
-    assert.deepEqual(requests.find((entry) => entry.method === "repo.task.create")?.executor, {
-      kind: "agent",
-      id: "runtime-cli-test",
-    });
-    assert.equal(
-      requests.some((entry) => entry.code === "invalid_request"),
-      false,
-      JSON.stringify(requests.filter((entry) => entry.code === "invalid_request")),
-    );
-    // A contracted read method with no CLI argv (repo.tasks.documents.list declares a closed payload
-    // without executor) stands in for "the next new command": with an explicit HARNESS_ACTOR, injection
-    // must follow the daemon-declared surface, so the real daemon accepts the request instead of
-    // rejecting an undeclared executor.
-    const probe = await runCommandThroughDaemon(
-      {
-        rootDir: safePath(root),
-        repoId: "runtime-cli",
-        json: true,
-        method: "repo.tasks.documents.list",
-        action: { kind: "task-documents-list", taskId },
-      },
-      undefined,
-      { env: { ...env, HARNESS_ACTOR: "agent:injection-probe" } },
-    );
-    assert.equal(probe.ok, true, JSON.stringify(probe));
-  } finally {
-    runMaybe(root, env, ["daemon", "stop"]);
-    rmSync(parent, { recursive: true, force: true });
-  }
-});
-
-async function runtimeInvariantEvidence(
-  root: string,
-  artifactRoot: string,
-  spawned: Record<string, unknown>,
-  settled: Record<string, unknown>,
-): Promise<Record<string, unknown>> {
-  const dispatchId = String(spawned.dispatchId),
-    runtimeSessionId = String(spawned.runtimeSessionId),
-    archivePath = path.join(artifactRoot, "dispatches", `${dispatchId}.json`),
-    reportPath = path.join(artifactRoot, "reports", `${dispatchId}.md`),
-    archive = JSON.parse(await readPublishedDispatch(archivePath)) as Record<string, unknown>,
-    events = makeTaskEventReader({ repoId: "runtime-cli", rootDir: root })
-      .read()
-      .events.filter(
-        (event) => "runtimeSessionId" in event.payload && event.payload.runtimeSessionId === runtimeSessionId,
-      ),
-    observed = events.filter((event) => event.type === "runtime_session_outcome_observed");
-  await eventuallyFile(reportPath);
-  return {
-    outcome: settled.outcome,
-    exitCode: ((settled.session as Record<string, unknown>).activity as Record<string, unknown>).exitCode,
-    archiveExists: existsSync(archivePath),
-    archiveOutcome: archive.outcome,
-    archiveExitCode: archive.exitCode,
-    reportExists: existsSync(reportPath),
-    exitedEvents: events.filter((event) => event.type === "runtime_session_exited").length,
-    observedEvents: observed.length,
-    observedOutcome: observed[0]?.type === "runtime_session_outcome_observed" ? observed[0].payload.outcome : null,
-    observedExitCode: observed[0]?.type === "runtime_session_outcome_observed" ? observed[0].payload.exitCode : null,
-  };
-}
-function readDispatchRecords(root: string, dispatchId: string): Record<string, unknown>[] {
-  return readFileSync(path.join(root, ".harness", "runtime", "dispatches", `${dispatchId}.jsonl`), "utf8")
-    .trim()
-    .split(/\r?\n/u)
-    .filter(Boolean)
-    .map((line) => JSON.parse(line) as Record<string, unknown>);
-}
-async function eventuallyNotification(
-  root: string,
-  dispatchId: string,
-): Promise<{ readonly started: Record<string, unknown> | undefined; readonly finished: Record<string, unknown> }> {
-  let records: Record<string, unknown>[] = [];
-  for (let attempt = 0; attempt < 500; attempt += 1) {
-    records = readDispatchRecords(root, dispatchId).filter((record) => record.kind === "exit_notification");
-    const finished = records.find((record) => record.phase === "finished");
-    if (finished) return { started: records.find((record) => record.phase === "started"), finished };
-    await new Promise((resolve) => setTimeout(resolve, 20));
-  }
-  throw new Error(`notification did not settle: ${JSON.stringify(records)}`);
-}
-async function readPublishedDispatch(target: string): Promise<string> {
-  await eventuallyFile(target);
-  return readFileSync(target, "utf8");
-}
-async function eventuallyFile(target: string): Promise<void> {
-  for (let attempt = 0; attempt < 500; attempt += 1) {
-    if (existsSync(target)) return;
-    await new Promise((resolve) => setTimeout(resolve, 20));
-  }
-  throw new Error(`file did not appear: ${target}`);
-}
-async function eventually<T>(read: () => T): Promise<T> {
-  let last: T | undefined;
-  for (let attempt = 0; attempt < 100; attempt += 1) {
-    last = read();
-    const rows = (last as Record<string, unknown>).dispatches;
-    if (Array.isArray(rows) && rows.some((row) => (row as Record<string, unknown>).status === "cancelled")) return last;
-    await new Promise((resolve) => setTimeout(resolve, 20));
-  }
-  throw new Error(`dispatch did not settle: ${JSON.stringify(last)}`);
-}
-async function eventuallyRuntimeStatus(
-  root: string,
-  env: NodeJS.ProcessEnv,
-  runtimeSessionId: string,
-  liveness: string,
-): Promise<{ readonly session: Record<string, unknown> & { readonly activity: Record<string, unknown> } }> {
-  let last: Record<string, unknown> = {};
-  for (let attempt = 0; attempt < 500; attempt += 1) {
-    last = run(root, env, ["runtime", "status", runtimeSessionId]);
-    const session = last.session as Record<string, unknown> & { readonly activity: Record<string, unknown> };
-    if (session.liveness === liveness) return { session };
-    await new Promise((resolve) => setTimeout(resolve, 20));
-  }
-  throw new Error(`runtime status did not reach ${liveness}: ${JSON.stringify(last)}`);
-}
-function processAlive(pid: number): boolean {
-  if (!Number.isInteger(pid) || pid < 1) return false;
-  try {
-    process.kill(pid, 0);
-    return true;
-  } catch {
-    return false;
-  }
-}
-function run(root: string, env: NodeJS.ProcessEnv, args: readonly string[]): Record<string, unknown> {
-  const result = runMaybe(root, env, args);
-  assert.equal(result.status, 0, `${result.stderr}\n${JSON.stringify(result.receipt)}`);
-  return result.receipt;
-}
-// Writes return at durable acceptance; a test that reads Git or the worktree waits for both followers explicitly.
-function published(root: string, env: NodeJS.ProcessEnv, receipt: Record<string, unknown>): Record<string, unknown> {
-  const wait = ["--wait", "git_verified,worktree_visible", "--timeout-ms", "5000"];
-  return run(root, env, ["receipt", "show", String(receipt.opId), ...wait]);
-}
-function runMaybe(
-  root: string,
-  env: NodeJS.ProcessEnv,
-  args: readonly string[],
-): {
-  readonly status: number | null;
-  readonly pid: number | undefined;
-  readonly receipt: Record<string, unknown>;
-  readonly stderr: string;
-} {
-  const result = spawnSync(process.execPath, [cli, "--root", root, "--json", ...args], { encoding: "utf8", env });
-  return {
-    status: result.status,
-    pid: result.pid,
-    receipt: result.stdout.trim() ? (JSON.parse(result.stdout) as Record<string, unknown>) : {},
-    stderr: result.stderr,
-  };
-}
-function runAsync(
-  root: string,
-  env: NodeJS.ProcessEnv,
-  args: readonly string[],
-): Promise<{
-  readonly status: number | null;
-  readonly pid: number | undefined;
-  readonly receipt: Record<string, unknown>;
-  readonly stderr: string;
-}> {
-  return new Promise((resolve, reject) => {
-    const child = spawn(process.execPath, [cli, "--root", root, "--json", ...args], {
-      env,
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-    let stdout = "",
-      stderr = "";
-    child.stdout.setEncoding("utf8").on("data", (chunk: string) => (stdout += chunk));
-    child.stderr.setEncoding("utf8").on("data", (chunk: string) => (stderr += chunk));
-    child.once("error", reject);
-    child.once("close", (status) =>
-      resolve({
-        status,
-        pid: child.pid,
-        receipt: stdout.trim() ? (JSON.parse(stdout) as Record<string, unknown>) : {},
-        stderr,
-      }),
-    );
-  });
-}
-async function eventuallyRuntimeReaderReuse(
-  userRoot: string,
-  daemonId: string,
-): Promise<{ readonly conn: string; readonly helloCount: number; readonly statusReads: number }> {
-  let observed: Record<string, string[]> = {};
-  for (let attempt = 0; attempt < 100; attempt += 1) {
-    observed = {};
-    const logRoot = path.join(userRoot, "logs"),
-      names = existsSync(logRoot)
-        ? readdirSync(logRoot).filter((name) => name.startsWith(`daemon-${daemonId}-conn-`) && name.endsWith(".jsonl"))
-        : [];
-    for (const name of names)
-      for (const line of readFileSync(path.join(logRoot, name), "utf8").trim().split("\n").filter(Boolean)) {
-        const record = JSON.parse(line) as Record<string, unknown>;
-        if (record.event !== "request" || typeof record.conn !== "string" || typeof record.method !== "string")
-          continue;
-        const key = `${String(record.pid)}:${record.conn}`;
-        (observed[key] ??= []).push(record.method);
-      }
-    const reused = Object.entries(observed).find(
-      ([, methods]) =>
-        methods.filter((method) => method === "repo.agentRuntime.sessions.read").length >= 2 &&
-        methods.filter((method) => method === "protocol.hello").length === 1,
-    );
-    if (reused)
-      return {
-        conn: reused[0],
-        helloCount: 1,
-        statusReads: reused[1].filter((method) => method === "repo.agentRuntime.sessions.read").length,
-      };
-    await new Promise((resolve) => setTimeout(resolve, 20));
-  }
-  throw new Error(`runtime status wait did not reuse one daemon connection: ${JSON.stringify(observed)}`);
-}
-function assertTaskMissionPrompt(
-  prompt: string,
-  expected: {
-    readonly repoId: string;
-    readonly taskId: string;
-    readonly canonicalRoot: string;
-    readonly workerRoot: string;
-    readonly taskPackageRoot: string;
-    readonly daemonUserRoot: string;
-    readonly daemonId: string;
-    readonly runtimeSessionId: string;
-    readonly mission: string;
-  },
-): void {
-  assert.ok(
-    prompt.includes(
-      `# Dispatch Preconditions\nRepository id: ${expected.repoId}\nRepository registration: enabled\nCanonical repository root: ${expected.canonicalRoot}\nWorker repository root: ${expected.workerRoot}\nCanonical Task ID: ${expected.taskId}\nTask package root: ${expected.taskPackageRoot}\nDaemon user root: ${expected.daemonUserRoot}\nDaemon id: ${expected.daemonId}\nDaemon endpoint: `,
-    ),
-    prompt,
-  );
-  assert.ok(prompt.includes(`\nRuntime actor: agent:runtime-session:${expected.runtimeSessionId}\n`), prompt);
-  assert.ok(prompt.endsWith(`# Assigned Mission\n${expected.mission}`), prompt);
-}
-function writeIdentity(target: string, identity: Record<string, unknown>): void {
-  const kind = String(identity.kind) as "agent" | "squad",
-    declaration = kind === "agent" ? identity.agent : identity.squad;
-  mkdirSync(target, { recursive: true });
-  writeFileSync(
-    path.join(target, kind === "agent" ? "agent.json" : "squad.json"),
-    `${JSON.stringify({ schema: `${kind}-declaration/v1`, ...declaration }, null, 2)}\n`,
-  );
-}
-function writeProvider(target: string, version: string): void {
-  writeProviderExecutable(
-    target,
-    `const fs = require("node:fs");\nconst args = process.argv.slice(2);\nif (args[0] === "--version") { console.log("codex ${version}"); process.exit(0); }\nif (args[0] === "login" && args[1] === "status") process.exit(0);\nconst prompt = fs.readFileSync(0, "utf8"), mission = prompt.split("\\n# Assigned Mission\\n").at(-1), secret = "sk-runtime-secret-1234567890";\nif (mission === "failure:empty") process.exit(1);\nif (mission === "failure:secret") process.stderr.write("OPENAI_API_KEY=" + secret + "\\n", () => process.exit(1));\nelse if (mission === "failure:structured") process.stdout.write([JSON.stringify({ type: "thread.started", thread_id: "provider-cli-session" }), JSON.stringify({ type: "turn.failed", error: { message: "structured provider failure", apiToken: secret } })].join("\\n") + "\\n", () => process.exit(1));\nelse { const resumed = args[0] === "exec" && args[1] === "resume", readOnly = mission === "read-only", noAction = mission === "no-action";\nif (readOnly && !args.includes("read-only")) process.exit(9);\nconst session = resumed ? args.at(-2) : "provider-cli-session";\nconst batch = mission.includes("batch hold"), mark = (event) => { if (batch) fs.appendFileSync(".batch-tracker", event + "\\n"); };\nconst emit = () => { mark("start"); console.log(JSON.stringify({ type: "thread.started", thread_id: session })); if (readOnly) console.log(JSON.stringify({ type: "item.completed", item: { id: "inspect", type: "command_execution", command: "ls packages", aggregated_output: "cli daemon kernel", exit_code: 0, status: "completed" } })); console.log(JSON.stringify({ type: "item.completed", item: { id: "live", type: "agent_message", text: "live:" + prompt } })); if (mission === "hold") setInterval(() => {}, 1000); else { if (!readOnly && !noAction) console.log(JSON.stringify({ type: "item.completed", item: { id: "write", type: "file_change", changes: [{ path: "result.txt", kind: "add" }], status: "completed" } })); console.log(JSON.stringify({ type: "item.completed", item: { id: "final", type: "agent_message", text: resumed ? "resumed:" + session + ":" + prompt : "final:" + prompt } })); console.log(JSON.stringify({ type: "turn.completed", usage: { input_tokens: 1, output_tokens: 1 } })); mark("end"); } };\nif (mission === "stream prompt") setTimeout(emit, 3000); else if (batch) setTimeout(emit, 250); else emit(); }\n`,
-  );
-}
-function writeProgressProvider(target: string, version: string): void {
-  writeProvider(target, version);
-  const source = readFileSync(target, "utf8"),
-    batchMarker =
-      'const batch = mission.includes("batch hold"), mark = (event) => { if (batch) fs.appendFileSync(".batch-tracker", event + "\\n"); };\n',
-    threadMarker = 'console.log(JSON.stringify({ type: "thread.started", thread_id: session })); if (readOnly)',
-    quotaMarker = "else { const resumed",
-    quotaFailure =
-      'else if (mission === "failure:429") process.stdout.write([JSON.stringify({ type: "thread.started", thread_id: "provider-cli-session" }), JSON.stringify({ type: "turn.failed", error: { http_status: 429, code: "insufficient_quota", message: "credit balance exhausted", reset_at: "2026-09-06T05:06:07Z" } })].join("\\n") + "\\n", () => process.exit(1));\nelse { const resumed',
-    progressSetup =
-      `${batchMarker}const progressTask = mission.startsWith("progress-middle:") ? mission.slice("progress-middle:".length) : null;\n` +
-      'const submitTask = mission.startsWith("submit-before-exit:") ? mission.slice("submit-before-exit:".length).split(":")[0] : null;\n',
-    progressWrite =
-      `console.log(JSON.stringify({ type: "thread.started", thread_id: session })); if (progressTask) { for (const text of ["Provider checkpoint one.", "Provider checkpoint two."]) { let result; for (let attempt = 0; attempt < 20; attempt += 1) { result = require("node:child_process").spawnSync(process.execPath, [${JSON.stringify(cli)}, "--root", process.cwd(), "--json", "task", "progress", "append", progressTask, "--text", text, "--evidence", "test:reports/runtime-progress.txt:provider checkpoint"], { encoding: "utf8", env: process.env }); if (result.status === 0) break; Atomics.wait(new Int32Array(new SharedArrayBuffer(4)), 0, 0, 50); } if (result.status !== 0) { process.stderr.write("progress append failed: " + result.stdout + result.stderr); process.exit(8); } } } ` +
-      `if (submitTask) { const packageRoot = prompt.split("Task package root: ")[1].split("\\n")[0]; fs.writeFileSync(require("node:path").join(packageRoot, "closeout.md"), "# Closeout\\n\\n## Summary\\n\\nRuntime worker submitted before exit at " + mission.split(":").at(-1) + ".\\n\\n## Verification\\n\\nIntegration runtime submission.\\n\\n## Residual Risk\\n\\nNone.\\n\\n## Same Mechanism Elsewhere\\n\\nRuntime archive lifecycle.\\n"); const result = require("node:child_process").spawnSync(process.execPath, [${JSON.stringify(cli)}, "--root", process.cwd(), "--json", "task", "submit", submitTask], { encoding: "utf8", env: process.env }); if (result.status !== 0) { process.stderr.write("task submit failed: " + result.stdout + result.stderr); process.exit(8); } } if (readOnly)`,
-    next = source
-      .replace(batchMarker, progressSetup)
-      .replace(threadMarker, progressWrite)
-      .replace(quotaMarker, quotaFailure);
-  if (next === source) throw new Error("runtime progress provider marker changed");
-  writeFileSync(target, next);
-}
